troubleshoots and prepares testing
Ammends various sort function to accept a list of dictionaries, and a key (found in all dictionaries).  Sort functions then sort the whole list of dictionaries, as if they are rows in a dataframe based on the key (~dataframe column name).  Also adds time testing and time testing results capture